Move news admin tasks and update imports
