+++ handlers/news/tasks_register.go
@@ -4,8 +4,6 @@ import "github.com/arran4/goa4web/internal/tasks"
 
 // RegisterTasks registers news related tasks with the global registry.
 func RegisterTasks() {
-	tasks.Register(NewsUserAllowTask)
-	tasks.Register(NewsUserRemoveTask)
 	tasks.Register(announcementAddTask)
 	tasks.Register(announcementDeleteTask)
 	tasks.Register(editReplyTask)
